Updated landing page Banner component
