+++ src/assets/styles/Banner.css
@@ -1,0 +1,42 @@
+/* Fading animation
+.elementToFadeInAndOut {
+  background: none;
+  -webkit-animation: fadeinout 4s linear forwards;
+  animation: fadeinout 4s linear forwards;
+}
+
+@-webkit-keyframes fadeinout {
+  0%,
+  100% {
+    opacity: 0;
+  }
+  50% {
+    opacity: 1;
+  }
+}
+
+@keyframes fadeinout {
+  0%,
+  100% {
+    opacity: 0;
+  }
+  50% {
+    opacity: 1;
+  }
+} */
+
+.slide1 {
+  background-image: url('../images/home-bg-1.jpg');
+}
+
+.slide2 {
+  background-image: url('../images/home-bg-2.jpg');
+}
+
+.slide3 {
+  background-image: url('../images/home-bg-4.jpg');
+}
+
+.swiper-pagination-bullet-active {
+  background-color: rgb(60, 241, 90) !important;
+}
