+++ src/components/Banner.jsx
@@ -1,77 +1,83 @@
-import bgImg from "../assets/images/slide-5.png";
-import bgImg1 from "../assets/images/slide-6.png";
-import bgImg2 from "../assets/images/slide-7.png";
-import bgImg3 from "../assets/images/categories_2.png";
-import { useEffect } from "react";
-import "./banner.css";
+import { useNavigate } from 'react-router-dom';
+
+import '../assets/styles/Banner.css';
+import { Swiper, SwiperSlide } from 'swiper/react';
+import { Pagination } from 'swiper';
+import 'swiper/css';
+import 'swiper/css/pagination';
 
 const Banner = () => {
-  var myIndex = 0;
-  useEffect(carousel);
-
-  function carousel() {
-    var i;
-    var x = document.getElementsByClassName("carousel-item");
-    for (i = 0; i < x.length; i++) {
-      x[i].style.display = "none";
-    }
-    myIndex++;
-    if (myIndex > x.length) {
-      myIndex = 1;
-    }
-    x[myIndex - 1].style.display = "flex";
-    setTimeout(carousel, 4000); // Change image every 2 seconds
-  }
+  const navigate = useNavigate();
 
   return (
-    <div className="py-36">
-      <div className="flex justify-center items-center ml-64">
-        {/* padding makes the container bigger */}
-        <div className="p-16">
-          <h1 className="text-6xl text-gray-800 font-medium mb-4 capitalize">
-            best raw materials for your business.
-          </h1>
-          <p>
-            Lorem ipsum dolor sit amet consectetur adipisicing elit. Deleniti ad
-            quam sapiente, neque quis pariatur.
-          </p>
-          <div className="mt-12">
-            <a
-              href="#"
-              className="bg-primary border border-primary text-white px-8 py-3 font-medium rounded-md hover:bg-transparent hover:text-primary transition"
-            >
-              Shop Now
-            </a>
+    <>
+      <span className="h-3 w-3 rounded-full bg-sky-500 bannerBullet"></span>
+      <Swiper
+        modules={[Pagination]}
+        pagination={{
+          clickable: true,
+        }}
+        slidesPerView={1}
+        spaceBetween={20}
+        className="w-full h-[85vh]"
+        loop={true}
+      >
+        <SwiperSlide>
+          <div className="w-full h-full bg-center bg-cover bg-no-repeat slide1">
+            <div className="w-1/2 h-full flex flex-col justify-center items-center gap-10">
+              <div className="w-96 font-bold text-6xl leading-[60px] text-white">
+                <span>Best materials for the best medicines</span>
+              </div>
+              <div className="w-96 h-14 px-10">
+                <button
+                  onClick={() => navigate('/products')}
+                  className="h-full w-full border-2 rounded-lg border-white font-bold text-2xl text-white hover:bg-white hover:text-sky-500 transition duration-200"
+                >
+                  Shop Now
+                </button>
+              </div>
+            </div>
           </div>
-        </div>
-        <div className="container">
-          <div
-            id="slide1"
-            className="carousel-item relative w-full justify-center"
-          >
-            <img src={bgImg} className="object-cover elementToFadeInAndOut" />
+        </SwiperSlide>
+        <SwiperSlide>
+          <div className="w-full h-full bg-center bg-cover bg-no-repeat slide2">
+            <div className="w-1/2 h-full flex flex-col justify-center items-center">
+              <div className="w-full font-bold flex justify-center gap-2 text-5xl brightness-125 mb-4">
+                <span className="text-gray-50">Medicine.</span>
+                <span className="text-sky-700">Reinvented.</span>
+              </div>
+              <div className="w-[400px] text-center text-gray-50 text-lg font-semibold mb-8">
+                <span>We're revolutionizing the pharmacy industry and we would like you to be a part of it</span>
+              </div>
+              <div className="w-96 h-14 px-10">
+                <button className="h-full w-full border-2 rounded-lg border-white font-bold text-2xl text-white hover:bg-white hover:text-sky-500 transition duration-200">
+                  Discover
+                </button>
+              </div>
+            </div>
           </div>
-          <div
-            id="slide2"
-            className="carousel-item relative w-full justify-center"
-          >
-            <img src={bgImg2} className="object-cover elementToFadeInAndOut" />
+        </SwiperSlide>
+        <SwiperSlide>
+          <div className="w-full h-full flex flex-col justify-center items-center gap-12 bg-center bg-cover bg-no-repeat slide3">
+            <div className="w-full flex justify-center text-5xl font-bold text-gray-50">
+              <span>The 6 type of vaccines technologies</span>
+            </div>
+            <div className="w-2/3 flex justify-center text-justify text-lg font-semibold text-white">
+              <span>
+                Ever since the first vaccine was developed in 1796 to treat smallpox, several different methods have been created to develop
+                successful vaccines. Today, those methods, known as vaccine technologies, are more advanced and use the latest technology to
+                help protect the world from preventable diseases. Do you know all of them?
+              </span>
+            </div>
+            <div className="w-96 h-14 px-10">
+              <button className="h-full w-full rounded-lg font-bold text-2xl bg-sky-500 text-white hover:bg-white hover:text-sky-500 transition duration-200">
+                Read More
+              </button>
+            </div>
           </div>
-          <div
-            id="slide3"
-            className="carousel-item relative w-full justify-center"
-          >
-            <img src={bgImg3} className="object-cover elementToFadeInAndOut" />
-          </div>
-          <div
-            id="slide4"
-            className="carousel-item relative w-full justify-center"
-          >
-            <img src={bgImg1} className="object-cover elementToFadeInAndOut" />
-          </div>
-        </div>
-      </div>
-    </div>
+        </SwiperSlide>
+      </Swiper>
+    </>
   );
 };
